core/requirements.txt
# general modules
django<4.3,>4.2
python-decouple
Pillow

# third party modules
djangorestframework
markdown
django-filter
drf-yasg[validation]
djangorestframework-simplejwt
django-mail-templated
# deployment modules
